requirements.txt
# Grasp - AI Educational Tutor
# Python 3.11+ recommended

# Core Framework
fastapi==0.115.0
uvicorn==0.31.0
python-dotenv==1.0.1
pydantic==2.9.2

# LangChain & Agents
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-groq==0.2.1
langgraph==0.2.45

# Database
neo4j==5.25.0
langchain-neo4j==0.1.0

# Search
tavily-python==0.3.3

# Data Processing
numpy==1.26.4
pandas==2.2.2

# PDF Processing
pymupdf
python-multipart

# Evaluation
ragas

# Observability (optional)
# arize-phoenix
# mlflow
